Don't duplicate MonoDroid.Tuner steps
Instead use ifdefs in the existing steps.
